Expose Extractor as part of Application
Restructure Extractor to its own directory.
