+++ src/blocks/extractor/textParser.js
