+++ src/scripts/build.js
@@ -1,5 +1,4 @@
 import { Application } from 'blocks/application';
-import { Extractor } from 'blocks/utilities/extractor';
 import { Logger } from 'blocks/utilities/logger';
 import { SitemapWriter } from 'blocks/writers/sitemap';
 import { FeedWriter } from 'blocks/writers/feed';
@@ -15,7 +14,7 @@ export const build = async () => {
   Logger.logProcessInfo();
 
   await Promise.all([
-    Extractor.extract().then(parsed => Application.initialize(parsed)),
+    Application.extractAndInitialize(),
     IconWriter.write(),
     ManifestWriter.write(),
   ]);
