+++ src/scripts/develop.js
@@ -1,5 +1,4 @@
 import { Application } from 'blocks/application';
-import { Extractor } from 'blocks/utilities/extractor';
 import { Logger } from 'blocks/utilities/logger';
 import { IconWriter } from 'blocks/writers/icon';
 import { AssetWriter } from 'blocks/writers/asset';
@@ -10,10 +9,7 @@ export const build = async () => {
   Logger.log('Build process is starting up...\n');
   Logger.logProcessInfo();
 
-  await Promise.all([
-    Extractor.extract().then(parsed => Application.initialize(parsed)),
-    IconWriter.write(),
-  ]);
+  await Promise.all([Application.extractAndInitialize(), IconWriter.write()]);
 
   await Promise.all([
     SearchIndexWriter.write(),
